refactored queries and client
hardly refactored all queries, changed dates from LocalDateTime to java.util.Date
client implementation now contains only part, related to fetching data from network

+++ src/main/java/org/openweathermap/api/model/Rain.java
@@ -9,5 +9,5 @@ public class Rain {
      * Rain volume for the last 3 hours
      */
     @SerializedName("3h")
-    private int threeHours;
+    private double threeHours;
 }
