+++ src/main/java/org/openweathermap/api/model/Snow.java
@@ -9,5 +9,5 @@ public class Snow {
      * Snow volume for the last 3 hours
      */
     @SerializedName("3h")
-    private int threeHours;
+    private double threeHours;
 }
